[FLINK-13290][hbase] Enable blink planner for integration tests of HBase
This commit combines HBaseTableSourceITCase and HBaseLookupFunctionITCase and HBaseConnectorITCase into one class.
This can save much cluster initialization time for us.
